home:s: tweak case study section.
I want to make the boxes in the Case study section the same size as the boxes in the next section.

I also want the boxes to be a color closer to the color of the background of that section. i want the color relatively the same as the black boxes below relative to their background.

We also should have the headings of those boxes (in case study) as same color as rest of section and for titles on same size as the numbers.